Fix bug where type of MembersInjector was always Unit (#238)
Previously the type of MembersInjector was always incorrectly set to unit,
this was due to the default value of the type being the returnType. In the case
of MembersInjector the type needed is that of the single argument.

This bug was previously surfacing as a duplicate binding error which
prevented any graph from having more than one MembersInjector.

Co-authored-by: Zac Sweers <pandanomic@gmail.com>